Basic CSS styling with inline-css and external CSS file
