+++ src/Components/Component.css
@@ -1,0 +1,4 @@
+.short-description {
+    color: green;
+    font-size: 20px;
+}
